refactor: extract everything but the concrete sections
