+++ src/schema/v2/homeView/HomeViewSectionTypeNames.ts
@@ -1,0 +1,20 @@
+export const HomeViewSectionTypeNames = {
+  HomeViewSectionActivity: "HomeViewSectionActivity",
+  HomeViewSectionArticles: "HomeViewSectionArticles",
+  HomeViewSectionArtists: "HomeViewSectionArtists",
+  HomeViewSectionArtworks: "HomeViewSectionArtworks",
+  HomeViewSectionAuctionResults: "HomeViewSectionAuctionResults",
+  HomeViewSectionCards: "HomeViewSectionCards",
+  HomeViewSectionDiscoverMarketingCollections:
+    "HomeViewSectionDiscoverMarketingCollections",
+  HomeViewSectionFairs: "HomeViewSectionFairs",
+  HomeViewSectionGalleries: "HomeViewSectionGalleries",
+  HomeViewSectionGeneric: "HomeViewSectionGeneric",
+  HomeViewSectionHeroUnits: "HomeViewSectionHeroUnits",
+  HomeViewSectionExploreByMarketingCollectionCategories:
+    "HomeViewSectionExploreByMarketingCollectionCategories",
+  HomeViewSectionMarketingCollections: "HomeViewSectionMarketingCollections",
+  HomeViewSectionSales: "HomeViewSectionSales",
+  HomeViewSectionShows: "HomeViewSectionShows",
+  HomeViewSectionViewingRooms: "HomeViewSectionViewingRooms",
+} as const
