+++ src/schema/v2/homeView/sections/DiscoverMarketingCollections.ts
@@ -1,7 +1,7 @@
 import { ContextModule, OwnerType } from "@artsy/cohesion"
 import { HomeViewSection } from "."
 import { withHomeViewTimeout } from "../helpers/withHomeViewTimeout"
-import { HomeViewSectionTypeNames } from "../HomeViewSection"
+import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
 import { connectionFromArray } from "graphql-relay"
 
 export const DiscoverMarketingCollections: HomeViewSection = {
