+++ src/schema/v2/homeView/sections/DiscoverSomethingNew.ts
@@ -1,6 +1,6 @@
 import { ContextModule } from "@artsy/cohesion"
 import { HomeViewSection } from "."
-import { HomeViewSectionTypeNames } from "../HomeViewSection"
+import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
 import { connectionFromArray } from "graphql-relay"
 
 const marketingCollectionSlugs = [
