+++ src/schema/v2/homeView/sections/RecommendedArtworks.ts
@@ -1,7 +1,7 @@
 import { ContextModule, OwnerType } from "@artsy/cohesion"
 import { HomeViewSection } from "."
 import { withHomeViewTimeout } from "../helpers/withHomeViewTimeout"
-import { HomeViewSectionTypeNames } from "../HomeViewSection"
+import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
 import { ArtworkRecommendations } from "schema/v2/me/artworkRecommendations"
 
 export const RecommendedArtworks: HomeViewSection = {
