+++ src/schema/v2/homeView/sections/ShowsForYou.ts
@@ -1,6 +1,6 @@
 import { ContextModule } from "@artsy/cohesion"
 import { HomeViewSection } from "."
-import { HomeViewSectionTypeNames } from "../HomeViewSection"
+import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
 
 export const ShowsForYou: HomeViewSection = {
   id: "home-view-section-shows-for-you",
