+++ src/schema/v2/homeView/sections/ViewingRooms.ts
@@ -1,6 +1,6 @@
 import { ContextModule, OwnerType } from "@artsy/cohesion"
 import { HomeViewSection } from "."
-import { HomeViewSectionTypeNames } from "../HomeViewSection"
+import { HomeViewSectionTypeNames } from "../HomeViewSectionTypeNames"
 
 export const ViewingRooms: HomeViewSection = {
   id: "home-view-section-viewing-rooms",
